chore(launcher): rename upgrade_server module to operation_server (file + main.rs)
